pyproject.toml
[project]
name = "trellis"
version = "1.0.0"
description = "Agentic pipeline platform — design agent teams that take ideas from concept to launch"
license = "Apache-2.0"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk",
    "nono-py",
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.0",
    "croniter",
    "typer[all]",
    "pyyaml",
    "rich",
    "watchfiles",
    "gitpython",
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "python-multipart",
    "websockets",
    "markdown",
    "anthropic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
trellis = "trellis.cli:app"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["trellis*"]

[tool.setuptools.package-data]
trellis = [
    "defaults/**/*",
    "defaults/**/.*",
    "web/frontend/templates/**/*",
    "web/frontend/static/**/*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
